🧽 clean up routes
